+++ tests/Support/CollectionTest.php
@@ -122,6 +122,16 @@ class CollectionTest extends TestCase
     }
 
     /** @test */
+    public function has()
+    {
+        $data = new Collection(['id' => 1, 'first' => 'Hello', 'second' => 'World']);
+        $this->assertTrue($data->has('first'));
+        $this->assertFalse($data->has('third'));
+        $this->assertTrue($data->has(['first', 'second']));
+        $this->assertFalse($data->has(['third', 'first']));
+    }
+
+    /** @test */
     public function each()
     {
         $c = new Collection($original = [1, 2, 'foo' => 'bar', 'bam' => 'baz']);