+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All Notable changes to `postmark-inbound` will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Added has method to Collection class.
+
 ### Fixed
 - Fixes tests to match expected/actual pattern. [`41b7910`](https://github.com/mvdnbrk/postmark-inbound/commit/41b791044a732b301588ea5de5d55a7f9990edf2)